reorganized code - rename base.py to app.py
